add whisper extension test

+++ intel_extension_for_transformers/llm/runtime/graph/scripts/ci/cpp_graph_inference.sh
@@ -231,6 +231,10 @@ function main() {
     elif [[ "${model}" == "magicoder" ]]; then
         quant_script="./build/bin/quant_llama"
         infer_cmd="./build/bin/run_llama"
+    elif [[ "${model}" == "whisper" ]]; then
+        quant_script="./build/bin/quant_whisper"
+        infer_cmd="./build/bin/run_whisper"
+        precision_list+=("q4_0")
     else
         echo "Error: Unexpedted model: $model" 1>&2
         exit 1
@@ -247,11 +251,13 @@ function main() {
     for p in "${precision_list[@]}"; do
         precisions_seen[$p]=x
     done
-    for p in "${extra_precision_list[@]}"; do
-        [[ ${precisions_seen[$p]} ]] && continue
-        precision_list+=("$p")
-        precisions_seen[$p]=x
-    done
+    if "${model}" != "whisper"; then
+        for p in "${extra_precision_list[@]}"; do
+            [[ ${precisions_seen[$p]} ]] && continue
+            precision_list+=("$p")
+            precisions_seen[$p]=x
+        done
+    fi
 
     # init conda
     #. $(dirname ${CONDA_EXE})/../etc/profile.d/conda.sh
@@ -355,18 +361,20 @@ function main() {
                     export LANG=en_US.UTF-8
                     export LC_ALL=en_US.UTF-8
                     echo "=======  Inference Start  ======="
+                    if "${model}" == "whisper";then OMP_NUM_THREADS=$cores_per_instance numactl -m 0 -C 0-$(($cores_per_instance - 1)) \
+                        $infer_cmd -f "/tf_dataset2/models/pytorch/whisper-tiny/jfk.wav"
+                    else
+                        real_ctx=$ctx # TODO(Zhenzhong): use same ctx for  chatglm & baichuan
+                        [[ "${model}" == "chatglm2" || "${model}" == "chatglm-6b" ||
+                            "${model}" == "baichuan-13b" || "${model}" == "baichuan2-13b" ]] && real_ctx=2047
 
-                    real_ctx=$ctx # TODO(Zhenzhong): use same ctx for  chatglm & baichuan
-                    [[ "${model}" == "chatglm2" || "${model}" == "chatglm-6b" ||
-                        "${model}" == "baichuan-13b" || "${model}" == "baichuan2-13b" ]] && real_ctx=2047
+                        OMP_NUM_THREADS=$cores_per_instance numactl -m 0 -C 0-$(($cores_per_instance - 1)) \
+                            $infer_cmd --seed 1234 -t $cores_per_instance -b 2047 -c $real_ctx -n ${output} -m ${model}-${precision}.bin $extension -p "$prompt" 2>&1 | tee ${WORKSPACE}/${logs_file} || true &
+                        monitor
 
-                    OMP_NUM_THREADS=$cores_per_instance numactl -m 0 -C 0-$(($cores_per_instance - 1)) \
-                        $infer_cmd --seed 1234 -t $cores_per_instance -b 2047 -c $real_ctx -n ${output} -m ${model}-${precision}.bin $extension -p "$prompt" 2>&1 | tee ${WORKSPACE}/${logs_file} || true &
-                    monitor
-
-                    echo "=======  Inference End  ======="
-                    python $script_dir/calculate_percentiles.py ${WORKSPACE}/${logs_file} ${model} ${precision} ${cores_per_instance} ${batch_size} ${input} ${output}
-
+                        echo "=======  Inference End  ======="
+                        python $script_dir/calculate_percentiles.py ${WORKSPACE}/${logs_file} ${model} ${precision} ${cores_per_instance} ${batch_size} ${input} ${output}
+                    fi
                     if [[ "$cores_per_instance" == "${cores_list[@]: -1:1}" ]] &&
                         [[ "$batch_size_idx" == "0" ]] &&
                         [[ "$input_idx" == "0" ]] &&
